Remove column F from Compatibility Matrix
Column F gave the maximum Java level supported for code being analyzed.
This was largely irrelevant, because your code could exceed this limit
without any problems. Only there would be no checks working with the
newer tokens. So the information was mostly useless or confusing.
The other columns move "up", so 'G' is the new 'F' ...