Publish fatJar, Eclipse Jar, and SonarQube Jar to Maven Central, too
GitHub does not allow '.jar' as an asset file type anymore, and we believe it's
asking too much to have everyone rename their downloaded artifacts. This changes
the file names of the Eclipse and SonarQube plugin Jars in order to comply with
Maven Central requirements. But at least they can stay '.jar'.